feat: 헤더 이슈 해결
