--- src/components/TopNavigation/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./TopNavigation"
